backend/requirements.txt
django==5.2.7
djangorestframework
djangorestframework-simplejwt
django-cors-headers
channels[daphne]
channels-redis
psycopg2-binary
gunicorn
python-dotenv